Fix in cluster
When comparing averageSMS to other SMS the relevant attributes of *all* particles were being using.
As a result SMS which differs only by their SM nodes were considered as different and clustering
was not performed.
As a solution, a new method (`getBSMattr`) was introduced for TheorySMS/AverageSMS, which
only consider BSM nodes.

 - Included testCluster
 - Included testConditions
 - Included testTheoryPredCombinations (requires pytorch to pass the test)